Data.GS1.DWhat

Documentation

data LabelEPC

Constructors

CL

Fields

_clClassLabelEpc :: ClassLabelEPC

_clQuantity :: Maybe Quantity

IL

Fields

_ilInstanceLabelEpc :: InstanceLabelEPC

newtype ParentLabel

Constructors

ParentLabel

Fields

unParentLabel :: InstanceLabelEPC

newtype InputEPC

Constructors

InputEPC

Fields

unInputEPC :: LabelEPC

newtype OutputEPC

Constructors

OutputEPC

Fields

unOutputEPC :: LabelEPC

data ObjectDWhat

Constructors

ObjectDWhat

Fields

_objAction :: Action

_objEpcList :: [LabelEPC]

data AggregationDWhat

Constructors

AggregationDWhat

Fields

_aggAction :: Action

_aggParentLabel :: Maybe ParentLabel

_aggChildEpcList :: [LabelEPC]

data TransactionDWhat

Constructors

TransactionDWhat

Fields

_transactionAction :: Action

_transactionParentLabel :: Maybe ParentLabel

_transactionBizTransactionList :: [BizTransaction]

_transactionEpcList :: [LabelEPC]

data TransformationDWhat

Constructors

TransformationDWhat

Fields

_transformationId :: Maybe TransformationId

_transformationInputEpcList :: [InputEPC]

_transformationOutputEpcList :: [OutputEPC]

data DWhat

The What dimension specifies what physical or digital objects participated in the event

Constructors

ObjWhat ObjectDWhat

AggWhat AggregationDWhat

TransactWhat TransactionDWhat

TransformWhat TransformationDWhat

readLabelEPC :: Maybe Quantity -> Text -> Either ParseFailure LabelEPC

Parses an EPC URN into a LabelEPC. As no quantity information is provided in a URN, Maybe Quantity is provided so ClassLabel EPCs can be instantiated with it.

urn2LabelEPC :: Text -> Either ParseFailure LabelEPC

Given an EPC URN, return a LabelEPC. As no quantity information is provided in the URN, the quantity field in ClassLabelEPC is set to Nothing. Use this when you do not need Quantity information