Data.GS1.EPC

Description

module containing error types, URI class, epc types the types in this file cover all dimensions

Documentation

class URI a where

Anything that could be converted into URI

Minimal complete definition

uriPrefix, uriSuffix, readURI

Methods

readURI :: Text -> Either ParseFailure a

renderURL :: a -> Text

Should not be directly implemented unless the format for URIs differs substantially from other URIs

data ParseFailure

Constructors

InvalidLength XMLSnippet
Length is not correct

InvalidFormat XMLSnippet
Components Missing, incorrectly structured, wrong payload

InvalidAction XMLSnippet
When parsing an action failed

InvalidBizTransaction XMLSnippet
When parsing a bizTransaction failed

InvalidEventId EventIdStr
When the `eventId` in the XML is not a valid UUID

TimeZoneError XMLSnippet
Error in parsing timezone

TagNotFound MissingTag
When a mandatory tag is not found

MultipleTags Text
When more than the specified number of tags are present

InvalidDispBizCombination XMLSnippet
When the disposition does not go with the bizstep

ChildFailure [ParseFailure]
When there is a list of Parsefailures typically applicable to higher level structures, like DWhat, DWhere, etc

newtype XMLSnippet

Constructors

XMLSnippet Text

newtype MissingTag

Constructors

MissingTag Text

newtype EventIdStr

Constructors

EventIdStr Text

newtype GS1CompanyPrefix

Assigned by a GS1 Member Organisation to a user/subscriber

Constructors

GS1CompanyPrefix

Fields

unGS1CompanyPrefix :: Text

newtype ItemReference

Constructors

ItemReference

Fields

unItemReference :: Text

newtype ExtensionDigit

Constructors

ExtensionDigit

Fields

unExtensionDigit :: Int

newtype SerialReference

Constructors

SerialReference

Fields

unSerialReference :: Text

newtype CheckDigit

Calculated according to an algorithm https://en.wikipedia.org/wiki/Global_Location_Number

Constructors

CheckDigit

Fields

unCheckDigit :: Int

newtype Lot

Constructors

Lot

Fields

unLot :: Text

newtype IndividualAssetReference

Constructors

IndividualAssetReference

Fields

unIndividualAssetReference :: Text

newtype SerialNumber

Constructors

SerialNumber

Fields

unSerialNumber :: Text

newtype SGLNExtension

Constructors

SGLNExtension

Fields

unSGLNExtension :: Text

newtype Uom

Constructors

Uom

Fields

unUom :: Text

newtype Amount

Constructors

Amount

Fields

unAmount :: Double

newtype AssetType

Constructors

AssetType

Fields

unAssetType :: Text

newtype Lng

Constructors

Lng

Fields

unLng :: Double

newtype Lat

Constructors

Lat

Fields

unLat :: Double

newtype LocationReference

Constructors

LocationReference

Fields

_locationRefVal :: Text

newtype DocumentType

Constructors

DocumentType

Fields

unDocumentType :: Text

newtype ServiceReference

Constructors

ServiceReference

Fields

unServiceReference :: Text

newtype BizTransactionId

Constructors

BizTransactionId

Fields

unBizTransactionId :: Text

newtype TransformationId

TransformationId From the spec EPCIS-Standard-1.2-r-2016-09-29.pdf Page 55 Some transformation business processes take place over a long period of time, and so it is more appropriate to represent them as a series of EPCIS events. A TransfomationId may be included in two or more TransformationEvents to link them together. When events share an identical TransformationId, the meaning is that the inputs to any of those events may have contributed in some way to each of the outputs in any of those same events.

Constructors

TransformationId

Fields

unTransformationId :: UUID

newtype EPCISTime

The TimeZone will be saved independently

Constructors

EPCISTime

Fields

unEPCISTime :: UTCTime

data SGTINFilterValue

Constructors

AllOthers

POSTradeItem

FullCaseForTransport

Reserved1

InnerPackTradeItemGroupingForHandling

Reserved2

UnitLoad

UnitInsideTradeItemOrComponentInsideAProductNotIntendedForIndividualSale

data Quantity

A quantity can be either the amount and the unit of measurement, or the item count.

Constructors

MeasuredQuantity

Fields

_quantityAmount :: Amount

_quantityUom :: Uom

ItemCount

Fields

_quantityCount :: Integer

data InstanceLabelEPC

Constructors

GIAI
Global Individual Asset Identifier, e.g. bucket for olives

Fields

_giaiCompanyPrefix :: GS1CompanyPrefix

_giaiSerialNum :: SerialNumber

SSCC
Serial Shipping Container Code

Fields

_ssccCompanyPrefix :: GS1CompanyPrefix

_ssccSerialNum :: SerialNumber

SGTIN
Serialsed Global Trade Item Number

Fields

_sgtinCompanyPrefix :: GS1CompanyPrefix

_sgtinSgtinFilterValue :: Maybe SGTINFilterValue

_sgtinItemReference :: ItemReference

_sgtinSerialNum :: SerialNumber

GRAI
Global Returnable Asset Identifier

Fields

_graiCompanyPrefix :: GS1CompanyPrefix

_graiAssetType :: AssetType

_graiSerialNum :: SerialNumber

data ClassLabelEPC

Constructors

LGTIN
LGTIN = GTIN + Batch/Lot scheme is used to denote a class of objects belonging to a given batch or lot of a given GTIN

Fields

_lgtinCompanyPrefix :: GS1CompanyPrefix

_lgtinItemReference :: ItemReference

_lgtinLot :: Lot

CSGTIN
Class SGTIN. e.g. olives in a vat, harvested in April 2017

Fields

_csgtinCompanyPrefix :: GS1CompanyPrefix

_csgtinSgtinFilterValue :: Maybe SGTINFilterValue

_csgtinItemReference :: ItemReference

data LocationEPC

EPCIS_Guideline.pdf Release 1.2 Ratified Feb 2017 - page 35 4.7.2 In EPCIS, the GLN+extension is represented as a Uniform Resource Identifier (URI) according to the EPC Tag Data Standard. The specific type of URI is called an SGLN, which is capable of representing either a GLN+extension or a GLN without extension. The SGLN for GLN 0614141111114 and extension 987 looks like this: urn:epc:id:sgln:0614141.11111.978

Constructors

SGLN

Fields

_sglnCompanyPrefix :: GS1CompanyPrefix

_locationRef :: LocationReference

_sglnExt :: Maybe SGLNExtension

data SourceDestType

EPCIS_Guideline.pdf Release 1.2 Ratified Feb 2017 - page 19 Source List and Destination List: is used to provide additional business context when an EPCIS event is part of a business transfer of ownership, responsibility or custody. As with business transactions, a source or destination is identified by a pair of identifiers: the type of the source or destination and an identifier of the source or destination of that type. The GS1 CBV (section 7.4.2) distinguishes three standard source/destination types: “owning_party”, “possessing_party”, “location”.

Constructors

SDOwningParty

SDPossessingParty

SDLocation

data BusinessTransactionEPC

Other GS1 object identifiers include GDTI for documents, GIAI for individual assets, GRAI for returnable assets, GSRN for services, GCN for coupons, and CPID for components or parts.

Constructors

GDTI

Fields

_gdtiCompanyPrefix :: GS1CompanyPrefix

_gdtiDocType :: DocumentType

_gdtiSerialNum :: SerialNumber

GSRN

Fields

_gsrnCompanyPrefix :: GS1CompanyPrefix

_gsrnSerialRef :: SerialReference

data LocationError

Constructors

IllegalGLNFormat

InvalidChecksum

data BizStep

CBV-Standard-1-2-r-2016-09-29.pdf Page 17

Constructors

Accepting

Arriving

Assembling

Collecting

Commissioning

Consigning

CreatingClassInstance

CycleCounting

Decommissioning

Departing

Destroying

Disassembling

Dispensing

Encoding

EnteringExiting

Holding

Inspecting

Installing

Killing

Loading

Other

Packing

Picking

Receiving

Removing

Repackaging

Repairing

Replacing

Reserving

RetailSelling

Shipping

StagingOutbound

StockTaking

Stocking

Storing

Transporting

Unloading

VoidShipping

data BizTransactionType

Constructors

Bol

Desadv

Inv

Pedigree

Po

Poc

Prodorder

Recadv

Rma

data BizTransaction

BizTransaction CBV Section 7.3 and Section 8.5

Constructors

BizTransaction

Fields

_btid :: BizTransactionId

_bt :: BizTransactionType

data Action

Constructors

Add

Observe

Delete

data Disposition

Constructors

Active

ContainerClosed

Damaged

Destroyed

Dispensed

Disposed

Encoded

Expired

InProgress

InTransit

Inactive

NoPedigreeMatch

NonSellableOther

PartiallyDispensed

Recalled

Reserved

RetailSold

Returned

SellableAccessible

SellableNotAccessible

Stolen

Unknown

readURI :: URI a => Text -> Either ParseFailure a

renderURL :: URI a => a -> Text

Should not be directly implemented unless the format for URIs differs substantially from other URIs

readURIClassLabelEPC :: [Text] -> Either ParseFailure ClassLabelEPC

readURIInstanceLabelEPC :: [Text] -> Either ParseFailure InstanceLabelEPC

mkAction :: Text -> Either ParseFailure Action

Orphan instances

Eq ZonedTime

Methods

(==) :: ZonedTime -> ZonedTime -> Bool

(/=) :: ZonedTime -> ZonedTime -> Bool

ToJSON TimeZone

Methods

toJSON :: TimeZone -> Value

toEncoding :: TimeZone -> Encoding

toJSONList :: [TimeZone] -> Value

toEncodingList :: [TimeZone] -> Encoding

FromJSON TimeZone

Methods

parseJSON :: Value -> Parser TimeZone

parseJSONList :: Value -> Parser [TimeZone]

ToSchema TimeZone

Methods

declareNamedSchema :: proxy TimeZone -> Declare (Definitions Schema) NamedSchema

ToParamSchema TimeZone

Methods

toParamSchema :: proxy TimeZone -> ParamSchema t