Data.GS1.Event

Documentation

data EventType

Constructors

ObjectEventT
When something is created

AggregationEventT
When something gets packaged into a container, or gets unpacked.

TransactionEventT
Some transaction has been made between 2 or more parties

TransformationEventT
Some product has been converted into another e.g A bucket of Tomatoes became some bottles of Tomato Sauce

data Event

Constructors

Event

Fields

_etype :: EventType

_eid :: Maybe EventId

_what :: DWhat

_when :: DWhen

_why :: DWhy

_where :: DWhere

allEventTypes :: [EventType]

getEventType :: DWhat -> EventType

Calls the appropriate stringify for a DWhat

mkEventType :: Text -> Maybe EventType

stringify :: IsString a => EventType -> a